docs(examples): update state example to async/await
